requirements.txt
requests>=2.31.0
PyGithub>=2.1.1
docker>=6.1.3
pyyaml>=6.0.1
gitpython>=3.1.40
jinja2>=3.1.2
python-dotenv>=1.0.0
anthropic>=0.40.0
